First pass at pinia workflow store; need to fix getters still
